+++ src/app/pages/forgot-password/forgot-password.module.ts
@@ -3,6 +3,7 @@ import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
 
+import { AuthStatusMessageComponent } from '../../components/auth-status-message/auth-status-message.component';
 import { ForgotPasswordPageRoutingModule } from './forgot-password-routing.module';
 import { ForgotPasswordPage } from './forgot-password.page';
 
@@ -14,6 +15,7 @@ import { ForgotPasswordPage } from './forgot-password.page';
     CommonModule,
     ReactiveFormsModule,
     IonicModule,
+    AuthStatusMessageComponent,
     ForgotPasswordPageRoutingModule
   ]
 })