+++ src/app/pages/login/login.module.ts
@@ -3,6 +3,7 @@ import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
 
+import { AuthStatusMessageComponent } from '../../components/auth-status-message/auth-status-message.component';
 import { LoginPageRoutingModule } from './login-routing.module';
 import { LoginPage } from './login.page';
 
@@ -14,6 +15,7 @@ import { LoginPage } from './login.page';
     CommonModule,
     ReactiveFormsModule,
     IonicModule,
+    AuthStatusMessageComponent,
     LoginPageRoutingModule
   ]
 })