+++ src/app/pages/reset-password/reset-password.module.ts
@@ -3,6 +3,7 @@ import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
 
+import { AuthStatusMessageComponent } from '../../components/auth-status-message/auth-status-message.component';
 import { ResetPasswordPageRoutingModule } from './reset-password-routing.module';
 import { ResetPasswordPage } from './reset-password.page';
 
@@ -14,6 +15,7 @@ import { ResetPasswordPage } from './reset-password.page';
     CommonModule,
     ReactiveFormsModule,
     IonicModule,
+    AuthStatusMessageComponent,
     ResetPasswordPageRoutingModule
   ]
 })