+++ src/app/pages/verify-email/verify-email.module.ts
@@ -2,6 +2,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { IonicModule } from '@ionic/angular';
 
+import { AuthStatusMessageComponent } from '../../components/auth-status-message/auth-status-message.component';
 import { VerifyEmailPageRoutingModule } from './verify-email-routing.module';
 import { VerifyEmailPage } from './verify-email.page';
 
@@ -12,6 +13,7 @@ import { VerifyEmailPage } from './verify-email.page';
   imports: [
     CommonModule,
     IonicModule,
+    AuthStatusMessageComponent,
     VerifyEmailPageRoutingModule
   ]
 })